Move the TraceConstraints into their own module
They pull in a lot of dependencies, and while the Protocol module does
need to refer to this bundle of constraints, it's not the natrual place
to define it.
